requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Marshmallow==0.15.0
Flask-Smorest==0.44.0
Authlib==1.3.1
Flask-Mail==0.9.1
marshmallow-sqlalchemy==0.29.0
requests>=2.31.0
marshmallow==3.21.1
python-dotenv==1.0.0
psycopg2-binary==2.9.10
Flask-Migrate==4.0.5
